+++ packages/arch3-core/src/signingarchwayclient.ts
@@ -22,7 +22,7 @@ import {
   HttpBatchClient,
   HttpBatchClientOptions,
   RpcClient,
-  Tendermint34Client,
+  Tendermint37Client,
   TendermintClient
 } from '@cosmjs/tendermint-rpc';
 import _ from 'lodash';
@@ -52,12 +52,12 @@ export interface SigningArchwayClientOptions extends SigningCosmWasmClientOption
    * Default adjustment factor to be multiplied against the estimate returned by the tx simulation.
    * If the gas limit is set manually in the transaction, this option is ignored.
    *
-   * @default 1.3
+   * @default 1.5
    */
   readonly gasAdjustment?: number;
 }
 
-const defaultGasAdjustment = 1.3;
+const defaultGasAdjustment = 1.5;
 
 interface DeliverTxResponseWithLogs extends DeliverTxResponse {
   readonly parsedLogs: readonly logs.Log[];
@@ -163,7 +163,7 @@ export class SigningArchwayClient extends SigningCosmWasmClient implements IArch
     signer: OfflineSigner,
     options: SigningArchwayClientOptions = {},
   ): Promise<SigningArchwayClient> {
-    const tmClient = await Tendermint34Client.connect(endpoint);
+    const tmClient = await Tendermint37Client.connect(endpoint);
     return SigningArchwayClient.createWithSigner(tmClient, signer, options);
   }
 
@@ -188,7 +188,7 @@ export class SigningArchwayClient extends SigningCosmWasmClient implements IArch
     batchClientOptions?: Partial<HttpBatchClientOptions>
   ): Promise<SigningArchwayClient> {
     const rpcClient: RpcClient = new HttpBatchClient(endpoint, batchClientOptions);
-    const tmClient = await Tendermint34Client.create(rpcClient);
+    const tmClient = await Tendermint37Client.create(rpcClient);
     return SigningArchwayClient.createWithSigner(tmClient, signer, options);
   }